+++ stories/formatted-input-currency.story.jsx
@@ -1,4 +1,4 @@
-import { number, select, text } from '@storybook/addon-knobs';
+import { boolean, number, select, text } from '@storybook/addon-knobs';
 import React from 'react';
 import { FormattedInputCurrency } from '../src';
 
@@ -16,6 +16,7 @@ const FormattedInputCurrencyStory = () => {
     decimals: number('Decimals', undefined),
     thousandSeparator: text('Thousand separator', undefined),
     decimalSeparator: text('Decimal separator', '.'),
+    selectValueOnClick: boolean('Select value on click', true),
   };
 
   const inputProps = {